feat: `test` feature, for now default
