+++ src/bgg/public_key.rs
@@ -82,6 +82,7 @@ impl<M: PolyMatrix> Evaluable for BggPublicKey<M> {
 }
 
 #[cfg(test)]
+#[cfg(feature = "test")]
 mod tests {
     use crate::{
         bgg::{circuit::PolyCircuit, sampler::BGGPublicKeySampler},
