+++ src/bgg/sampler.rs
@@ -166,6 +166,7 @@ where
 }
 
 #[cfg(test)]
+#[cfg(feature = "test")]
 mod tests {
     use super::*;
     use crate::{
