+++ src/poly/dcrt/params.rs
@@ -43,7 +43,7 @@ impl PolyParams for DCRTPolyParams {
     }
 }
 
-#[cfg(test)]
+#[cfg(feature = "test")]
 impl Default for DCRTPolyParams {
     fn default() -> Self {
         Self::new(4, 2, 17)
