+++ src/poly/dcrt/poly.rs
@@ -234,6 +234,7 @@ impl SubAssign<&DCRTPoly> for DCRTPoly {
 }
 
 #[cfg(test)]
+#[cfg(feature = "test")]
 mod tests {
     use super::*;
     use crate::poly::{dcrt::DCRTPolyUniformSampler, sampler::DistType, PolyParams};
