+++ src/poly/dcrt/sampler/hash.rs
@@ -161,6 +161,7 @@ where
 }
 
 #[cfg(test)]
+#[cfg(feature = "test")]
 mod tests {
     use super::*;
     use crate::poly::dcrt::DCRTPolyParams;
