+++ src/poly/dcrt/sampler/trapdoor.rs
@@ -223,6 +223,7 @@ impl PolyTrapdoorSampler for DCRTPolyTrapdoorSampler {
 }
 
 #[cfg(test)]
+#[cfg(feature = "test")]
 mod tests {
     use super::*;
     use crate::poly::{
